refactor: move TypeFlags enum to type utilities
